--- .dockerignore
@@ -1,5 +1,0 @@
-.*
-!.gitignore
-!.dockerignore
-
-node_modules
