+++ src/command/lib/commandsManual.js
