+++ src/command/lib/openai.js
